Release version 2.3.23
Source: 08cb6277006559db80b8d5a80aec9bc40389fbb1

Reviewed-on: https://gerrit.bouffalolab.com/c/release/bouffalo_sdk/+/9028
Tested-by: zrrong <zrrong@bouffalolab.com>

+++ drivers/soc/bl616cl/phyrf/include/wl_cli_api.h
@@ -5,8 +5,8 @@ typedef struct
 {
     uint32_t rxok;
     uint32_t rxend;
-    float ppm_dsss;
-    float ppm_ofdm;
+    int8_t ppm_dsss;
+    int8_t ppm_ofdm;
     int8_t rssi;
 } cli_ate_ret_t;
